Fix error when destination config.xml doesn't exist.
AFAICT the only reason this previously worked was that calling omero.cli.CLI() or self.cli.loadplugins() in Upgrade.setup_script_environment() auto-creates a default config.xml (see comment in Upgrade.__init__()).

Why has it broken now? No idea... maybe it's to do with importing multiple versions of the CLI in the nose test harness. Previously only one server was being upgraded in the test run.... perhaps when importing the CLI for the second time it isn't properly imported?